ivy_intelligence/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-allauth==0.63.6
celery==5.4.0
redis==5.0.8
django-celery-beat==2.7.0
requests==2.32.3
beautifulsoup4==4.12.3
scikit-learn==1.5.2
nltk==3.9.1
joblib==1.4.2
channels==4.1.0
daphne==4.1.2
Pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
